mirrors.bbclass: Add own copy of oe-core's bbclass
:Release Notes:
Remove catch-all entries which leak information about names and versions
of proprietary components, when the origin site isn't working or when
there is any other issue fetching it (e.g. typo in version, so that it
doesn't exist on primary site and all MIRRORs are checked after that)

:Detailed Notes:
Some entries need to be returned in order to fetch everything used in
build. And some entries needed to be removed e.g. GNUPG_MIRROR, because
this doesn't exist in Yocto 1.7 we're currently using.

:Testing Performed:
Build tested without premirror to see which entries we need to add
manually.

:QA Notes:

:Issues Addressed:
[PLAT-29114] Change MIRRORS variable so that upstream OE/YP mirrors
             aren't used for internal components
[PLAT-81460] CCC: mirrors.bbclass: Add own copy of oe-core's bbclass
[PLAT-81507] [OSE] CCC: Import few more changes

Change-Id: Ic27d6e3aa675ef5e839aac0abf338cc5e77be674