ls2-helpers: add temporary patch to fix build with gtest 1.8.0
:Release Notes:
This is only temporary change to unblock my builds with Yocto 2.3 Pyro which
contains gtest 1.8.0 instead of older 1.7.0 currently used in master.

It should be fixed correctly by ls2-helpers owner.

:Detailed Notes:
With gtest 1.7.0 we were changing this gtest behavior by:
meta-webos/recipes-upstreamable/gtest/gtest/0001-explicit-bool-cast.patch
but as it wasn't ever submitted to upstream, it might be easier
to fix our own users of gtest not to require this modification.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[PLAT-41150] Upgrade to Yocto 2.3 Pyro
[PLAT-78505] [OSE] CCC: Upgrade to Yocto 2.3 Pyro

Change-Id: Ice4b316be8dc525c15d7d7c969f16a8d02920a47