webos_filesystem_paths qml-webos-bridge=r11 (use empty QT_DIR_NAME variable)
:Release Notes:
We don't support qt4 for long time, maybe it's time to set QT_DIR_NAME
to empty. It's used in meta-qt5 only to allow both qt4 and qt5 in the
same build, we're not using that for very long time.

:Detailed Notes:
If we set QT_DIR_NAME we need all components to respect paths as set in
qmake5_paths and not to hardcode "qt5" suffix on their own add a work
around (symlink ${libdir}/qt5 > ${libdir}) in 5 components which
hardcode ${libdir}/qt5 in their QML files.

:Testing Performed:
Only build tested.

:QA Notes:
No functional change to the image.

:Issues Addressed:
[PLAT-43548] CCC: Use empty QT_DIR_NAME variable
[PLAT-48635] CCC: qt5: Upgrade to 5.6.3
[PLAT-62989] [OSE] CCC: qt5: Upgrade to 5.6.3 for webos

Change-Id: I90a9d4e549fba9bb80b3cb21bd0d94908658d04b
Originally-Reviewed-on: http://gpro.lgsvl.com/178192
Reviewed-on: http://gpro.lgsvl.com/208810
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Martin Jansa <martin.jansa@lge.com>
Tested-by: Martin Jansa <martin.jansa@lge.com>
Reviewed-by: Jaeyoon Jung <jaeyoon.jung@lge.com>
Reviewed-by: Kyungjik Min <dp.min@lge.com>