language-style=r2 node-inspector=r2 nodejs-module-webos-{dynaload=r5,pmlog=r5,sysbus=r8}
:Release Notes:
nodejs modules was shipping the node-gyp and npm to user home directory
changed to work directory of individual modules.

:Detailed Notes:
nodejs module was shipping the node-gyp and npm to home directory now it
is moved to work directories of individual modules by explicitly
exporting home and work directory.

:Testing Performed:
Testing done with Jackrabbit server & working fine
results are posted in PLAT-52171

:QA Notes:

:Issues Addressed:
[PLAT-52171] nodejs-module-webos-* fail to build on
             jackrabbit-build.lgsvl.net
[PLAT-52970] CCC: language-style=r2 node-inspector=r2
             nodejs-module-webos-{dynaload=r5,pmlog=r5,sysbus=r8}