webos_app.bbclass webos_app_generate_security_files.bbclass: improve manifest file generation for native apps.
:Release Notes:
Auto-Generate role files for native app
Add exeName in role files for native app
Generate manifest and role files for native app

:Detailed Notes:
Role files are not generated automatically for native apps/service
However as they will not run under container. In this case
we need to add exeName in role file as well as change the way
allowedName and other parameters are generated

Manifest file is not generated automatically for native apps/service
when we inherit webos_app script. This was happened due to setting
up WEBOS_SYSTEM_BUS_MANIFEST_TYPE which is not needed as webos_app
inturn inherits webos_app_generate_security_files

:Testing Performed:
Only build tested and verified in build that role files are
generated for native apps.

:QA Notes:

:Issues Addressed:
[PLAT-64095] [OSE] CCC: webos_app.bbclass
             webos_app_generate_security_files.bbclass: improve
             manifest file generation for native apps.
[PLAT-52025] Built-in Native app's ACG files do not  generated
             automatically at build time
[PLAT-60811] Built-in Native app's manifest file does not
             generated with webos_app