mesa=webos2
:Release Notes:
Fix leaking shaders attached with glUseProgram in mesa.

:Detailed Notes:
When glUseProgram is used, references to the included shaders are
added in ctx->Shader.ReferencedProgram. But those references are not
decreased when the shader data is deallocated. Thus, those shaders
are leaked.

Explicitely remove the pending references to these shaders.

:Testing Performed:
Launch/close cycles in webos, the shaders are not leaked anymore.

:QA Notes:
The right way to validate this is using valgrind, to verify we are not
leaking anymore the shader programs.

:Issues Addressed:
[PLAT-60137] CCC: mesa=webos2
[PLAT-51491] Memory leak is observed from WebAppMgr
[PLAT-52479] Upgrade to Yocto 2.5 Sumo
[PLAT-78511] [OSE] CCC: Upgrade to Yocto 2.5 Sumo

Change-Id: Ic2098add71bd3eccf90441ac5d4d5dad45e63c51