webos_enactjs_app.bbclass: improve manifest file generation for enactjs apps.
:Release Notes:
Generate manifest files for enactjs app

:Detailed Notes:
Manifest file is not generated automatically for enactjs apps
when we inherit webos_enactjs_app script. This was happened due to not setting
up WEBOS_SYSTEM_BUS_MANIFEST_TYPE

This was not needed for webos_app script. However since enactjs app
inherit webos_enactjs_app script it needs to be set here.

:Testing Performed:
Only build tested and verified in build that role files are
generated for native apps.

:QA Notes:

:Issues Addressed:
[PLAT-65305] [OSE ]CCC: webos_enactjs_app.bbclass:
              improve manifest file generation for enactjs apps
[PLAT-65620] [Web OS OSE] : No response from server in IOTIVITY SAMPLER.

Change-Id: Ifa754c71cb491120dbc175e1a53327bff008376c