rpi: rpi-config=webosrpi5 (resolve memory issues)
:Release Notes:
Resolve memory issues in RPi3 when running multiple apps in background.

:Detailed Notes:
We observed LSM crashes in RPi3 when we tried to run several(5-6) apps.
After some investigation it was clear that the issue is in Mesa driver.
There is an article in the RPi forum that says it is a well-known issue
and can be resolved with some configuration changes.
https://www.raspberrypi.org/forums/viewtopic.php?t=19334&start=125#p223549

To fix the issue it is necessary to remove the cma settings and add
gpu_mem settings instead.

:Testing Performed:
See the CCC JIRA ticket.

:QA Notes:
Test-cases: QO-4187

:Issues Addressed:
[PLAT-69055] Resolve issues after "fullscreen surface running in background" implementation
[PLAT-60829] LSM: Allow fullscreen surface running in background
[PLAT-78443] [OSE] CCC: rpi-config=webosrpi5 luna-surfacemanager=6

Change-Id: Ifc53f8fb27a560023c6158697882ea787a0cc3ef