connman=webos5
:Release Notes:
Fix for WiFi Network does not switch to the selected network
Fix "Unable to connect to the network" Popup appears instead
of Password prompt.

:Detailed Notes:
Whenever switching to different network, interface state will
be triggered with state DISCONNECT which disturbing the pending
wifi network connection status. Connecting to open network is a
side effect of the same. This patch applies a fix switch to
selected network.

:Testing Performed:
See Jira ticket PLAT-71526 for results.

:QA Notes:

:Issues Addressed:
[PLAT-71526] CCC: connman=webos5
[PLAT-57861] [webOS OSE] WiFi Connection :WiFi Network does not
             switch to the selected network.
[PLAT-65726] [webOS OSE] : "Unable to connect to the network" Popup
             appears instead of Password prompt.

Change-Id: Id58a9206c7094aed5fe036a84b46537bcccbf1b2