wam=r21
:Release Notes:
Remove dependency on oe.utils.conditional for sysbus files installation

:Detailed Notes:
wam: r20..r21
Due to our old bitbake version, oe.utils.conditional doesn't work
as expected always and caused issues with installation of sysbus files
for wam. We need the following patch on bitbake to fix
this issue properly as we use oe.utils.conditional in many places:
https://github.com/shr-distribution/bitbake/commit/e407d452d92956ea9d103ab60a3421d284fcff01

We had applied a workaround for missing feature in our bitbake version
by defining vardeps on variables that were made depend on oe.util.conditional()
However this workaround doesn't seem to be full-proof.
So, now we dont depend on oe.utils.conditional for installation of
sysbus files. We install them manually for both chr53 and chr68.

:Testing Performed:
Build checked for chr53 and chr68

:QA Notes:

:Issues Addressed:
[PLAT-71661] [webOS OSE]: Not able to launch all the Apps in OSE except web browser app
[PLAT-72195] [OSE] CCC: wam=r21

Change-Id: Id9e913c8d3197706148a0e81e1c22f85c62dc68d