wam=r27 (fix missing component in DEPENDS)
:Release Notes:
Add qt-features-webos in DEPENDS for wam.bb.

:Detailed Notes:
wam requires webos-services.prf provided by qt-features-webos in order
to install sysbus files. But it hasn't defined a build dependency to
that which results in sysbus files are not installed without an error if
wam is built prior to qt-features-webos.
WEBOS_SYSTEM_BUS_FILES_LOCATION has no effect and thus is removed if
WEBOS_SYSTEM_BUS_SKIP_DO_TASKS is set to 1.

:Testing Performed:
Build tested.

:QA Notes:

:Issues Addressed:
[PLAT-77897] [OSE] CCC: wam=r27 (fix missing component in DEPENDS)
[PLAT-75326] Home screen doesnt turn up while building PulseAudio in OSE
             CCC builds

Change-Id: I4998b5a20b3dec30828fcd8e38082a50decbc182