sam=r1 (migrate to Boost Signals2)
:Release Notes:
Deprecated signals library was dropped in 1.69.0 release:
https://www.boost.org/users/history/version_1_69_0.html

Signals (v1) is now removed. Its removal was announced in 1.68 and its
deprecation was announced in 1.54. Boost 1.68 is the last release that
provides this library. Users are encouraged to use Signals2 instead. The
Boost community thanks Douglas Gregor for his work on Signals which
served its users well and which also inspired Signals2.

This component is already using Signals2 in some places, migrate the
rest to use Signals2 as well.

:Detailed Notes:

:Testing Performed:
CCC testing

:QA Notes:

:Issues Addressed:
[PLAT-78370] sam, notificationmgr, appinstalld, sensord: are using long
             deprecated boost signals library
[PLAT-78507] [OSE] CCC: Upgrade to Yocto 2.6 Thud

Change-Id: I54ccc6590b603e1a01a367a35c1c8644fd3d61ba