notificationmgr=r7 (remove signals check from CMakeLists)
:Release Notes:
Deprecated signals library was dropped in 1.69.0 release:
https://www.boost.org/users/history/version_1_69_0.html

Signals (v1) is now removed. Its removal was announced in 1.68 and its
deprecation was announced in 1.54. Boost 1.68 is the last release that
provides this library. Users are encouraged to use Signals2 instead. The
Boost community thanks Douglas Gregor for his work on Signals which
served its users well and which also inspired Signals2.

This component is already using Signals2 so just drop the check for old
unused component.

:Detailed Notes:

:Testing Performed:
CCC testing

:QA Notes:

:Issues Addressed:
[PLAT-78370] sam, notificationmgr, appinstalld, sensord: are using long
             deprecated boost signals library
[PLAT-78507] [OSE] CCC: Upgrade to Yocto 2.6 Thud

Change-Id: I1a123852f771bb5a7f219532e8c379f9b76a91c6