webos.inc: Move WEBOS_DISTRO_PRERELEASE, WEBOS_TARGET_MACHINE_IMPL, WEBOS_TARGET_MACHINE_VARIANT variables to bbclass
:Release Notes:
Move these 3 variables:
WEBOS_DISTRO_PRERELEASE ?= "devel"
WEBOS_TARGET_MACHINE_IMPL ?= "emulator"
WEBOS_TARGET_MACHINE_VARIANT ?= "none"
to bbclasses which should be inherited when these variables are being
used. This way the default value is always defined when parsing the
recipes (even without webos as DISTRO) and vice versa, when someone
forgets to inherit the right bbclass the value won't be defined which
is easier to detect.

:Detailed Notes:
Remove webos_arch_indep from cmake-modules-webos-native as it doesn't
make sense with native recipe.

Add webos_prerelease_dep inherit to webos_lttng.bbclass and enable it
for prerelease builds automatically, otherwise it will fail to parse,
because the inner if expression will stay unexpanded, because of
undefined ${WEBOS_DISTRO_PRERELEASE}:
  bb.data_smart.ExpansionError: Failure expanding variable DEPENDS,
  expression was pkgconfig-native cmake-native  ${@icecc_dep_prepend(d)}
  virtual/arm-webos-linux-gnueabi-gcc
  virtual/arm-webos-linux-gnueabi-compilerlibs virtual/libc
  luna-service2 libpbnjson pmloglib glib-2.0 gtest
  cmake-modules-webos-native${@ ' lttng-ust' if '${@'0' if
  '${WEBOS_DISTRO_PRERELEASE}' == '' else '1'}' == '1' else ''}
  ninja-native which triggered exception SyntaxError: invalid syntax
  (DEPENDS, line 1)

For webos_cmake.bbclass we still need to define some default values
to make sure that the if expression is expanded, but at least define
it to something which shouldn't work and should help noticing the
missing inherit.

There is still one unexpanded variable in EXTRA_OECMAKE (and other
variables like that) PACKAGECONFIG_CONFARGS, but it's not used in any
expression, so it doesn't break anything (until someone defines the
same variable in shell env), I've sent patch to oe-core upstream:
http://lists.openembedded.org/pipermail/openembedded-core/2019-May/282782.html

:Testing Performed:

:QA Notes:

:Issues Addressed:
[PLAT-79498] [OSE] CCC: Fix metadata parsing without webos as a DISTRO

Change-Id: Icb7394b3c593f98394165a716d8990c3cee263a9