qtdeclarative=webos38 (fix pixmap cache issue in multiple windows)
:Release Notes:
Add QQmlEngine to QQuickPixmapKey.

:Detailed Notes:
A QQuickPixmapData created from the same image URL, size and transform
is meant to be shared if it is cached. However it should not be when it
is being used in multiple windows.
This patch adds the QQmlEngine to the key of the pixmap data cache so
that the cache hits only when it is created from the same engine.

Tagged as 5.6.meta-webos.38.

:Testing Performed:
Went through steps to reproduce in PLAT-80766.
Also confirmed by a debugging tool that the cache hits only when it
belongs to the same QML engine.

:QA Notes:

:Issues Addressed:
[PLAT-80925] [OSE] CCC: qtdeclarative=webos38 luna-surfacemanager=8
             qml-webos-framework=5 qtwayland=7 qtwayland-webos=6

Change-Id: I7155b9acbb04a867794d514c993201770557ac2c
Reviewed-on: http://gpro.lge.com/246093
Reviewed-by: Martin Jansa <martin.jansa@lge.com>
Tested-by: Martin Jansa <martin.jansa@lge.com>
Reviewed-by: Jongwoong Lee <jw.lee@lge.com>
Reviewed-by: Minjae Kim <nate.kim@lge.com>