webos-sota.inc: enable FOTA by using ostree
:Release Notes:
Enable FOTA by using ostree (meta-updater)

:Detailed Notes:
The ostree (by meta-updater) is an incremental FOTA solution
that offers compact update image size than A/B based FOTA.

Becasue of the meta-updater designed to use 'usrmerge' DISTRO_FEATURE,
the rootfs directory structure will be changed.
By this, the /usr contains most things to update system.

Following directories will be moved and linked as symbolic.
 - /sbin => /usr/sbin
 - /bin => /usr/bin
 - /lib => /usr/lib
 - /home => /var/rootdirs/home => /sysroot/home
 - /mnt => /var/rootdifs/mnt

Following directories will be mounted as bind.
 - /var => /sysroot/ostree/deploy/webos/var/
 - /media => /sysroot/media

By default with meta-updater, the /usr are mounted as 'read only'
and it makes bad development environment for most users.
So, we enables the 'hotfix' mode of ostree to make /usr as 'read/write'.

Also, because of 'usrmerge' changes root_prefix,
we need to use proper variable to prepend the install directory.
This patch changes many component's install stage.
 - crashd: r10..r11
 - ecryptfs-utils: r0..r1
 - iotivity-node: r2..r3
 - webos_enactjs_app.bbclass
 - webruntime: r21..r22
 - com.webos.service.contextintentmgr: ..webosupdater1
 - ntfs-3g-ntfsprogs: ..webosupdater1
 - systemtap: ..webosupdater1
 - webos-initscript: ..webosupdater1

Some changes on the meta-webos-raspberrypi was copied and touched from
https://github.com/advancedtelematic/meta-updater-raspberrypi

And, patches are added to u-boot because of the u-boot
(v2018.07, and also recent) doesn't support rpi4 yet.

For more detail, please visit following URL.
 - https://wiki.lgsvl.com/display/webOSDocs/libostree+for+webOS+OSE

This solution will be ignored when target is 'qemux86'.

:Testing Performed:

:QA Notes:

:Issues Addressed:
[PLAT-76088] [FOTA] Prepare meta layer for libostree
[PLAT-85195] [FOTA] Add u-boot rpi4 support for ostree
[PLAT-81123] CCC: webos-sota.inc: enable FOTA by using ostree

Change-Id: Icccb4c0fda39fb5b24d37619e99e38fb1eb6e037