qtdeclarative=webos42
:Release Notes:
Fix an issue where an item does not receive a pointer event if there is
a child item that accepts the event even when that child item is shown
under the parent item as per the stacking order.

:Detailed Notes:
qtdeclarative=webos42
Add 0026-QQuickWindow-Consider-z-order-of-children-when-deliv.patch

Tagged as 5.12.meta-webos.11.

:Testing Performed:
See the CCC JIRA ticket.
Tested steps in PLAT-83233.

:QA Notes:

:Issues Addressed:
[PLAT-84636] CCC: qtdeclarative=webos42
[PLAT-83233] [GMRSI] Touch event isn't delivered to web engine when
             user touches video area

Change-Id: I2a7f4e039cb82be97a9628dadb25d3e73824e448