webos_cmake.bbclass: sync from cmake-modules-webos-native _webos_set_from_env
:Release Notes:
As the comment above this section says, these variables are exported for
_webos_set_from_env in cmake-modules-webos-native, remove the variables
which don't belong there.

:Detailed Notes:
Remove entries for:
webos_eacg_certificates
webos_key_dir
webos_trust_level

webos_eacg_certificates doesn't seem to be needed exported, it's used
only by eacg-keys.bb. webos_key_dir, webos_trust_level are used only for
build time signing in opkg-utils, but this isn't used in webOS and if
it needs to be exported, then it should be exported everywhere not only
in recipes inheriting webos_cmake.

This adds webos_testsdir to make sure that the move from /opt to
/usr/opt is respected by cmake recipes (e.g. luna-service2).

:Testing Performed:
Only build tested.

:QA Notes:

:Issues Addressed:
[PLAT-85921] Move /opt into /usr/opt
[PLAT-85926] CCC: Move /opt into /usr/opt

Change-Id: Iba31e3c29feb8c891e6bc9a7e59f349ecbc6e942