qtdeclarative=webos45
:Release Notes:
Patch was excluded in order to prevent unexpected text behaviour

:Detailed Notes:
As described in PLAT-87434 text starts to marquee when tile is focused.
We found out that the root cause in the QtDeclarative.
We compared it with pure qtdeclarative 5.12.3. As result we found out
that the function QQuickTextPrivate::setupTextLayout works
not as properly because of layoutInProgress flag.
In qml KeywordComponent -> ScrollTextBox have incorrect
value of truncated and incorrect behaviour when
scrollingText.horizontalAlignment executs. We suggest to exclude patch 20.
Also we checked issue PLAT-63024 with our equipment
the issue doesn`t reproduce.

5.12.meta-webos.17

:Testing Performed:
Master 1845

1. Press Home button.
2. Click on the *Search* tile in the system tiles on the left side of
the launcher ribbon.
3. Type some request and hit the "Search" button
4.Repeat step 4 several times
5.Press Home button.
6.Move the focus to the *Search* tile in the launcher ribbon.
7.Press Up key and move the focus to each tile and wait a bit.

:QA Notes:

:Issues Addressed:
[PLAT-87434] Text starts to marquee when search tile list is focused
[PLAT-95034] CCC: qtdeclarative=webos45

Change-Id: Ie0740f4a54b14dca47d0245cce6f6f3583467df0