webos-recipe-blacklist.inc: blacklist s3c24xx-gpio, s3c64xx-gpio, wmiconfig, usbpath
:Release Notes:
All these are failing because of svn fetcher issue,
svn://svn.openmoko.org is now automatically redirected to
https://github.com/openmoko/openmoko-svn where it's
fetched, but into different directory, S does not point to correct
directory and build fails.

:Detailed Notes:
Reported upstream in:
http://lists.openembedded.org/pipermail/openembedded-devel/2020-February/205012.html
then fixed in dunfell:
https://github.com/openembedded/meta-openembedded/commit/6902dcd2edc3a08f6929e19e9a9c18493883f9d2
and then all of these were removed:
https://github.com/openembedded/meta-openembedded/commit/7449a3e08389a0c50cdf17bc16995ef1678dc2d2
so these PNBLACKLISTs can also be removed when we're upgrading to
dunfell.

:Testing Performed:

:QA Notes:

:Issues Addressed:
[PLAT-109540] CCC: Various build fixes

Change-Id: I84c753956e453e5a8140179b4725cec29d78969d