qtbase=webos79webosrpi12 qtdeclarative=webos53 webos-initscripts=r15
:Release Notes:
Remove unnecessary patches.

:Detailed Notes:
Remove patches which confirmed by us and Qt team as unnecessary,
and will not apply to upstream.

Redesign qttestability.so loading system - use Qt env. variable
Redesign some patches as Qt team suggested.

qtbase=webos79
(-) WebOS-platform-expects-filenames-in-UTF-8.patch
(-) Fix-HarfBuzz-NG-regression.patch
(-) Check-if-combined-glyph-exists-in-font-s-charmap-tab.patch
(-) Fix-bug-of-calculate-xsize-for-fonts.patch
(-) Move-testability-loading-code-to-the-QGuiApplication.patch

qtbase=webosrpi12
Refresh patches with rebased to the patches in meta-webos.

qtdeclarative=webos53
(-) LTTNG-tracing-support-in-Qt-Quick.patch
(-) Fix-incorrect-listview-footer-position-on-empty-mode.patch
(-) QQuickItemView-Remove-redundant-refill-when-layout.patch
(m) Set-the-stencil-buffer-zone.patch
(m) QQuickWindow-Consider-z-order-of-children-when-deliv.patch

webos-initscipts=r15
Setup WEBOS_QTTESTABILITY_ENABLED environment variable according
release version of build.

:Testing Performed:
Please ref. to GT-31669

:QA Notes:

:Issues Addressed:
[PLAT-113540] CCC: Remove unnecessary Qt patches

Change-Id: I3d79abcd1efcb63485ce02f99bfd7390adff8ec5