webos_qmllint.bbclass: Improve as per Qt 6
:Release Notes:
Improve webos_qmllint.bbclass as per changes with Qt 6.

:Detailed Notes:
Skip empty files to avoid unnecessary error from qmllint.
Drop comments before running the pattern matching.
Check the old syntax for signal connection in Connections. It is treated
as a warning since the old syntax is allowed as of the current Qt
version. If WEBOS_QMLLINT_ERROR_ON_WARNING is set to 1, it will fail if
there is any matching case.
Add patterns not allowed in Qt 6. "url" type properties with relative
URLs are not allowed. Inline shader code for "fragmentShader" or
"vertexShader" property is not allowed.

:Testing Performed:
Tested with an empty .qml file.
Tested with WEBOS_QMLLINT_ERROR_ON_WARNING set to 1.
Tested files with patterns above.

:QA Notes:
N/A

:Issues Addressed:
[PLAT-143667] CCC: Upgrade to Qt 6 for OSE and Auto
[PLAT-129932] Detect old QML syntax at build time

Change-Id: Ibc38dcfc5548c047f1188c796135f8edb85e49fe