packagegroup-webos-extended: remove media-codec-interface
:Release Notes:
media-codec-interface.bb has for some reason this restriction:

COMPATIBLE_MACHINE = "^qemux86$|^raspberrypi3$|^raspberrypi3-64$|^raspberrypi4$|^raspberrypi4-64$"

so it cannot be added in RDEPENDS for all MACHINEs.

:Detailed Notes:
And it probably isn't needed at all in the image unless
neva-webrtc PACKAGECONFIG is enabled in webruntime, so
remove it here and let webruntime pull it into the image
only when enabled.

:Testing Performed:

:QA Notes:
NA

:Issues Addressed:
[PLAT-136606] CCC: ose: Add media-codec-interface=1-r0,v1.0.0 webruntime=r46 webruntime_84=13
[PLAT-135803] Implement video encoder interface for webos
[PLAT-135561] CCC: Upgrade to latest revisions from Yocto 3.1 Dunfell

Change-Id: I5112829c82d4eb148c13140f60aecc5c92bfa0cb