webos.inc: add QT_VERSION variable
:Release Notes:
Add QT_VERSION to handle differences among meta-qtX version.

:Detailed Notes:
QT_VERSION is introduced to handle differences among meta-qtX
version. For example, qtshadertools-native is N/A in meta-qt5 but it is
required for some recipes with meta-qt6.

DEPENDS_append = " ${@ 'qtshadertools-native' if d.getVar('QT_VERSION', True) == '6' else '' }"

This variable would be used to handle such cases.

:Testing Performed:
Build webos and webos-auto with meta-qt6

:QA Notes:
N/A

:Issues Addressed:
[PLAT-143667] CCC: Upgrade to Qt 6 for OSE and Auto
[PLAT-139717]

Change-Id: I69a008e2a7a4c7a1e7e23352ac924731359fd54c