webos: qtbase=webosrpi0
:Release Notes:
Drop qtbase_git.bbappend in meta-webos-raspberrypi.

:Detailed Notes:
Patches don't apply to Qt 6. They should be re-written for Qt 6 if they
are still needed. However I don't think so since builds for raspberrypi3
haven't been managed for a long time and the patch for raspberrypi4 is
not needed as the cursor is not used in OSE as per the current UX.

:Testing Performed:
See the CCC JIRA ticket.

:QA Notes:
N/A

:Issues Addressed:
[PLAT-143667] CCC: Upgrade to Qt 6 for OSE and Auto

Change-Id: Ife96921a826eb5abe7758221a1c46d6785ad3ae5