webruntime=r49 (Add exportbuildscript.inc to export a build script)
:Release Notes:
Add exportbuildscript.inc to export a build script on SDK environment.

:Detailed Notes:
Building webruntime out of oe-build is required to speed up of building
webruntime especially when upgrading previous models because it takes
too much time to build whole dependencies of the webruntime.

This patch was made by proposing build webruntime on SDK environment so
that we don't need to prepare all the dependencies of the webruntime
because SDK has already all packages installed.

Only required things are preparing default GN_ARGS variable from the
current webOS build and customizing the variables for the target SDK.

'exportbuildscript' task is added after do_configure before do_compile
to generate a build script and the build script contains GN_ARGS which
is used from the webOS build executing and it allows to override
several GN_ARGS for the the target SDK.

The script is doing mainly
 - The build script automatically sets TARGET_CPU based on the
   OECORE_TARGET_ARCH variable of the SDK environment.
   TARGET_CPU variable is used for gn args such as target_cpu and
   v8_snapshot_toolchain.
 - It detects arm architecture from TUNE_FEATURES variable of the SDK
   environment and sets arm architecture specific gn args.
 - Allows to override or substitute a specific argument
   It overrides tool chain and sysroot related variables by default to
   build in SDK.
 - Allows to remove a specific argument

:Testing Performed:
Build succeed and build.sh is generated on the ${B} directory

:QA Notes:
N/A

:Issues Addressed:
[WRN-1740] Implement chromium standalone build using OE configuration
[WRN-5579] CCC: webruntime=r49

Change-Id: I36c78abada7b9d3f58d65d512857e9f8a1418c11