qtbase=webos92 qtdeclarative=webos72
:Release Notes:
Update patches as per the meta-qt6 layerpin update.

:Detailed Notes:
qtbase=webos92
Add 0002-Revert-QRegularExpression-fix-matching-over-null-emp.patch for
the qtbase-native compile issue.
Drop 0001-Fix-button-type-to-update-coordinates-when-dragging.patch as
it is included in the updated layerpin.

qtdeclarative=webos72
Drop 0001-qqmllistmodel-Fix-QObjects-setting-indestructible.patch as it
is included in the updated layerpin.

:Testing Performed:
See the CCC JIRA ticket.

:QA Notes:
N/A

:Issues Addressed:
Refer to the CCC JIRA ticket for related issues.
[WRN-6965] CCC: Update meta-qt6 to v6.2.0-beta4

Change-Id: I041d94adb24d3746bc0787dcca35e61c31f264e2