libtool=webos1 (patch to fix libtool filters sanitizer flags)
:Release Notes:
From https://savannah.gnu.org/patch/?8775 founded by pulseaudio build

:Detailed Notes:
from: https://wiki.gentoo.org/wiki/AddressSanitizer/Problems
When linking shared libraries libtool will filter the LDFLAGS
and only allows known good flags. The current version of libtool is
unaware of Address Sanitizer and will therefore also filter
the -fsanitize=address flag from the linker call. This breaks the
compilation.

This has been fixed upstream and this fix is part of the libtool
ebuild in the asantoo overlay.
This fix is not part of a libtool release yet.
It will take a while until this change takes effect.
The problem is that the libtool scripts are bundled with applications.
Therefore even if upstream libtool allows the -fsanitize=address flag,
this change has to be made on the systems of the upstream packagers.

As the number of packages affected by this is huge the asantoo-base
package in the overlay installs a Portage hook that will
automatically patch all applications that ship
an affected libtool script (the problem is in the ltmain.sh file).

:Testing Performed:
refer to CCC

:Issues Addressed:


Change-Id: I881ffbc2a01ca4f7d1ced78da63aacb7b08dc1fc