sstate.bbclass : To fix errors about read-only sstate mirrors
:Release Notes:
Fix errors about read-only sstate mirrors
This patch should be removed when yocto version up.

:Detailed Notes:
https://git.openembedded.org/openembedded-core
/commit/?h=dunfell&id=f2360f796016cce93610a080ff07c8047bda6ee8
If a read-only sstate mirror is used in conjunction with hash equiv,
then OSError will be raised when an sstate-cache hit is achieved.

This is because sstate_task_postfunc will try to "touch" the symlinks
that point to the read-only sstate mirror when sstate_report_unihash
has changed SSTATE_PKG.

:Testing Performed:
build tested

:QA Notes:
No changes to image

:Issues Addressed:

Change-Id: I5098b7632406bc9a1db40a439931d86892f138ec