qtbase=webos97 qtwayland=webos27 (apply patches conditionally)
:Release Notes:
Apply patches conditionally so that they don't conflict with a newer
Qt version.
Fix a crash when touch device is removed.

:Detailed Notes:
qtbase=webos97
Add 0002-evdev-Prevent-race-condition-in-touch-events-process.patch
Apply the patch below if PV is less than the version where it is merged.
0001-Support-to-get-timing-from-pagefilp.patch
0002-evdev-Prevent-race-condition-in-touch-events-process.patch

qtwayland=webos27
Apply the patch below if PV is less than the version where it is merged.
0001-Support-presentation-time-protocol.patch
Apply the patch below with no condition as it hasn't been merged yet.
0002-Use-scope-resolution-operator-for-request.patch

I tried to use 'maxrev' option in SRC_URI but found that it doesn't work
as documented. The implementation isn't correct as it compares commit
hashes as string instead of the order in the commit history.
'maxver' is the option not documented, but works as expected.

:Testing Performed:
See the CCC JIRA ticket.

:QA Notes:
N/A

:Issues Addressed:
Refer to the CCC JIRA ticket for related issues.
[WRN-15812] CCC: luna-surfacemanager=364 qtwayland-webos=72 and more

Change-Id: Ic641e5c307f1811a27a342f4bc787eb6f49b294c