wayland: update patches to apply on 1.20.0 version
:Release Notes:

:Detailed Notes:
The last patch needs to be reworked a bit to use meson instead
of autotools after:
https://gitlab.freedesktop.org/wayland/wayland/-/commit/3bda3d1b4729c8ee7c533520a199611cb841bc8f

and there were quite a few conflicts from:
https://gitlab.freedesktop.org/wayland/wayland/-/commit/ca893075ef75f2ad4688d8fc1dcc20260f91afe9

When rebasing this to 1.21.0 version I've noticed:
https://gitlab.freedesktop.org/wayland/wayland/-/commit/ca4678354c4a7c21bfd184e604ad7a3c8224e10c
which is useful for 1.20.0 and allows to simplify
this patch quite a bit while allowing to use almost
the same patch (with small conflict in meson.build)
for 1.21.0 as well.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRO-12999] CCC: Upgrade to Yocto 4.0 Kirkstone
[WRN-9820] Create GPVB with Yocto 4.0 Kirkstone
[WRO-12543] [OSE-RPI4]: Error output "No such file or directory"

Change-Id: Ic76d1322570983e8321cd5811193ec44d595494c