meta-webos-backports-4.0: import as-is in meta-oe/kirkstone
:Release Notes:
The purpose of meta-webos-backports layers is to be able to drop the backported
recipes without unexpected side-effects, any webOS specific changes should still
be implemented in .bbappend, so that they stay even after the backports layer
is removed.

:Detailed Notes:
The only acceptable exception is the inherit of distutils3-base instead of
setuptools3-base in opencv, as setuptools3-base doesn't exist yet in dunfell
and distutils3-base works well enough in this case.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRO-4984] CCC: Various build fixes

Change-Id: I1285c29e3fcdb285a8df667b53ea2eb872c7ec7b