gcc-source=webos3 (use standard Upstream-Status in .patch files)
:Release Notes:
Add the version from which these changes were backported
to make it easier in following changes which upgrade to newer
Yocto with gcc-12.1 (through 10.2 in gatesgarth, 10.3 in hardknot
11.2 in honister, ..)

:Detailed Notes:
Replace 0007-arm-xnnpack-neondot-build-error-fix.patch with individual
patches from upstream releases/gcc-9 branch, so that they can be
separately dropped when upgrading to newer gcc which includes some of
them.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRO-5487] CCC: tensorflow-lite=r0,v2.6.2 edgeai-vision=7-r0,v1.0.0
[WRO-4984] CCC: Various build fixes

Change-Id: I0d73c239404792760d379155a130272e503fa534