libedgetpu: move one directory up for bitbake to actually find the recipe
:Release Notes:
It's not included in any image by default and nobody noticed
that it's not even included in BBFILES because of extra directory
level not included in:
BBFILES += "${LAYERDIR}/recipes-*/*/*.bb ${LAYERDIR}/recipes-*/*/*.bbappend"

:Detailed Notes:
Fixes:
$ bitbake libedgetpu
ERROR: Nothing PROVIDES 'libedgetpu'. Close matches:
  libdnet
  libedit

but the recipe is still broken:
| /bin/bash: line 1: bazel: command not found

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRO-4984] CCC: Various build fixes

Change-Id: I48a9722a04b0f32de7fee9bf229e1c203dad3c4b