ose: sam=64
:Release Notes:
Fix timeout for retrying killing running applications

:Detailed Notes:
The GLib function used to set up a timeout to wait for closing or
killing applications, g_timeout_add, accepts an "interval" parameter
in milliseconds.
But, when an application is closed, the close transition timeout is
set to 1 (1ms), when it was actually expected to wait 1 second per
attempt to retry closing/killing an application.Fix the timeout
(and define a constant to store the value).

sam: submissions/63..submissions/64
445570c Fix timeout for retrying killing running applications

:Testing Performed:
 Test Result updated in WRO-8090

:QA Notes:
 N/A

:Issues Addressed:
[WRO-8090] CCC : sam=64
[WRO-7198][RDX_CRASH][webos] /usr/bin/WebAppMgr in __poll (libc.so.6 + 0xc78e4)

Change-Id: Id1991e072765cb8a5c6aaf99aa14b52af31cc2bd