ose: g-media-pipeline=21.gav.33
:Release Notes:
Handle Invalid Argument exception thrown from stoll

:Detailed Notes:
start parameter is parsed and the next string is passed to stoll.
If it is not a valid string with a number content, it throws an invalid
argument error and program crashes. The exception is caught to avoid the
crash
g-media-pipeline:
submissions/21.gav.32...submissions/21.gav.33
262413b Handle Invalid Argument exception for stoll

:Testing Performed:
Test results are updated in WRO-9605

:QA Notes:
NA

:Issues Addressed:
[WRO-9605] CCC: g-media-pipeline=21.gav.33
[WRO-9224] [RDX_CRASH][webos] /usr/sbin/g-media-pipeline in raise

Change-Id: I383a4c85850699e49ebf2185ffc334bd0771d5f1