com.webos.app.videocall: temporarily enable network in do_compile task
:Release Notes:
bitbake can now restrict network access in various tasks based on "network" flag as implemented in:
  https://git.openembedded.org/bitbake/commit/?id=0746b6a2a32fec4c18bf1a52b1454ca4c04bf543
  https://git.openembedded.org/bitbake/commit/?id=9d6341df611a1725090444f6f8eb0244aed08213
and oe now enables network mostly only in do_fetch and few special tasks as implemented in:
  https://git.openembedded.org/openembedded-core/commit/?id=7ce1e88a3ad85bbb925bb9f7167dc0a5fd1c27f4

:Detailed Notes:
This needs to be fixed properly in the recipes, this work around is meant only
temporary.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRO-12999] CCC: Upgrade to Yocto 4.0 Kirkstone
[WRN-9820] Create GPVB with Yocto 4.0 Kirkstone

Change-Id: I5171a879570b58a65cdba9aef07da4028387730f