base-files=webos15 webos-initscripts=86
:Release Notes:
Set the db8datadir permission 755 rather than 700
Fix a prepare-system-dirs issue on emulator without serial setting

:Detailed Notes:
base-files: webos14..webos15
- Move db8datadir to 755

After we moved the unifiedsearch service as non-Root,
it's not possible to access /var/db/... on the emulator.

There is no problem on the RPI4 because of the db directory
is created on runtime due to the ostree folder structure.

webos-initscripts: submissions/85..submissions/86
Fix the prepare-system-dirs service file

:Testing Performed:
Build and runtime tested

:QA Notes:
N/A

:Issues Addressed:
[WRO-10543] CCC: base-files=webos15 webos-initscripts=86
[WRO-10169] [OSE Emulator] Searched key app is not displaying in the terminal
[WRO-10194] [OSE Emulator] 1 loaded units listed is displaying.

Change-Id: Id600b860e826e56561e1422358be6493e3273969